changed the button block selectors; converted header test to hero [skip ci]
