+++ cypress/integration/v3/blocks/feature/index.js
@@ -46,8 +46,8 @@ function typeContent() {
 			.assertBlockContent( '.stk-block-heading__text', 'Feature block' )
 		cy.typeBlock( 'stackable/feature', '.stk-block-text__text', 'Lorem ipsum dolor sit amet.', 0 )
 			.assertBlockContent( '.stk-block-text__text', 'Lorem ipsum dolor sit amet.' )
-		cy.typeBlock( 'stackable/feature', '.stk-block-button__inner-text', 'Click here', 0 )
-			.assertBlockContent( '.stk-block-button__inner-text', 'Click here' )
+		cy.typeBlock( 'stackable/feature', '.stk-button__inner-text', 'Click here', 0 )
+			.assertBlockContent( '.stk-button__inner-text', 'Click here' )
 	} )
 }
